fix cast bug

+++ src/networks/transformer.py
@@ -324,7 +324,7 @@ class CustomSchedule(tf.keras.optimizers.schedules.LearningRateSchedule):
 
     def __call__(self, step):
         arg1 = tf.math.rsqrt(tf.cast(step, tf.float32))
-        arg2 = step * (self.warmup_steps ** -1.5)
+        arg2 = tf.cast(step, tf.float32) * (self.warmup_steps ** -1.5)
 
         return tf.math.rsqrt(self.d_model) * tf.math.minimum(arg1, arg2)
